# openapi.yaml
openapi: 3.1.0
info:
  title: API de Despacho y Logística --- Grupo 6
  description: Servicio de gestión de despachos v1.3
  version: 1.3.0
paths:
  /api/v1/shipments/quotes:
    post:
      summary: Quote Shipment
      operationId: quote_shipment_api_v1_shipments_quotes_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/shipments:
    post:
      summary: Create Shipment
      operationId: create_shipment_api_v1_shipments_post
      parameters:
      - name: x-request-id
        in: header
        required: true
        schema:
          type: string
          description: Identificador único de la petición
          title: X-Request-Id
        description: Identificador único de la petición
      - name: x-correlation-id
        in: header
        required: true
        schema:
          type: string
          description: Identificador para trazabilidad distribuida
          title: X-Correlation-Id
        description: Identificador para trazabilidad distribuida
      - name: x-consumer
        in: header
        required: true
        schema:
          type: string
          description: Identifica el cliente que consume el servicio
          title: X-Consumer
        description: Identifica el cliente que consume el servicio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShipmentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentCreateResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
    get:
      summary: Get Shipments
      operationId: get_shipments_api_v1_shipments_get
      parameters:
      - name: orderId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Orderid
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ShipmentStatus'
          - type: 'null'
          title: Status
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Pagesize
      - name: x-request-id
        in: header
        required: true
        schema:
          type: string
          description: Identificador único de la petición
          title: X-Request-Id
        description: Identificador único de la petición
      - name: x-correlation-id
        in: header
        required: true
        schema:
          type: string
          description: Identificador para trazabilidad distribuida
          title: X-Correlation-Id
        description: Identificador para trazabilidad distribuida
      - name: x-consumer
        in: header
        required: true
        schema:
          type: string
          description: Identifica el cliente que consume el servicio
          title: X-Consumer
        description: Identifica el cliente que consume el servicio
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ShipmentListResponse'
                - type: array
                  items:
                    $ref: '#/components/schemas/ShipmentResponse'
                title: Response Get Shipments Api V1 Shipments Get
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
  /api/v1/shipments/{shipment_id}:
    get:
      summary: Get Shipment By Id
      operationId: get_shipment_by_id_api_v1_shipments__shipment_id__get
      parameters:
      - name: shipment_id
        in: path
        required: true
        schema:
          type: string
          title: Shipment Id
      - name: x-request-id
        in: header
        required: true
        schema:
          type: string
          description: Identificador único de la petición
          title: X-Request-Id
        description: Identificador único de la petición
      - name: x-correlation-id
        in: header
        required: true
        schema:
          type: string
          description: Identificador para trazabilidad distribuida
          title: X-Correlation-Id
        description: Identificador para trazabilidad distribuida
      - name: x-consumer
        in: header
        required: true
        schema:
          type: string
          description: Identifica el cliente que consume el servicio
          title: X-Consumer
        description: Identifica el cliente que consume el servicio
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
    patch:
      summary: Update Shipment Status
      operationId: update_shipment_status_api_v1_shipments__shipment_id__patch
      parameters:
      - name: shipment_id
        in: path
        required: true
        schema:
          type: string
          title: Shipment Id
      - name: x-request-id
        in: header
        required: true
        schema:
          type: string
          description: Identificador único de la petición
          title: X-Request-Id
        description: Identificador único de la petición
      - name: x-correlation-id
        in: header
        required: true
        schema:
          type: string
          description: Identificador para trazabilidad distribuida
          title: X-Correlation-Id
        description: Identificador para trazabilidad distribuida
      - name: x-consumer
        in: header
        required: true
        schema:
          type: string
          description: Identifica el cliente que consume el servicio
          title: X-Consumer
        description: Identifica el cliente que consume el servicio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShipmentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentUpdateResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
  /api/v1/shipments/{shipment_id}/history:
    get:
      summary: Get Shipment History
      operationId: get_shipment_history_api_v1_shipments__shipment_id__history_get
      parameters:
      - name: shipment_id
        in: path
        required: true
        schema:
          type: string
          title: Shipment Id
      - name: x-request-id
        in: header
        required: true
        schema:
          type: string
          description: Identificador único de la petición
          title: X-Request-Id
        description: Identificador único de la petición
      - name: x-correlation-id
        in: header
        required: true
        schema:
          type: string
          description: Identificador para trazabilidad distribuida
          title: X-Correlation-Id
        description: Identificador para trazabilidad distribuida
      - name: x-consumer
        in: header
        required: true
        schema:
          type: string
          description: Identifica el cliente que consume el servicio
          title: X-Consumer
        description: Identifica el cliente que consume el servicio
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentHistoryResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          anyOf:
          - items:
              $ref: '#/components/schemas/ErrorDetail'
            type: array
          - type: 'null'
          title: Details
        correlationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlationid
      type: object
      required:
      - code
      - message
      title: ErrorResponse
...
    QuoteRequest:
      properties:
        city:
          type: string
          maxLength: 100
          title: City
        packages:
          items:
            $ref: '#/components/schemas/PackageInput'
          type: array
          title: Packages
      type: object
      required:
      - city
      - packages
      title: QuoteRequest
    QuoteResponse:
      properties:
        totalShippingCost:
          type: integer
          title: Totalshippingcost
        currency:
          type: string
          title: Currency
          default: CLP
      type: object
      required:
      - totalShippingCost
      title: QuoteResponse
    ShipmentCreate:
      properties:
        orderId:
          type: string
          title: Orderid
          description: ID único del pedido de G5
        customerName:
          type: string
          maxLength: 200
          title: Customername
          description: Nombre del destinatario
        address:
          type: string
          maxLength: 500
          title: Address
          description: Dirección física de entrega
        city:
          type: string
          maxLength: 100
          title: City
          description: Ciudad de destino
        packages:
          items:
            $ref: '#/components/schemas/PackageInput'
          type: array
          title: Packages
      type: object
      required:
      - orderId
      - customerName
      - address
      - city
      - packages
      title: ShipmentCreate
    ShipmentCreateResponse:
      properties:
        shipmentIds:
          items:
            type: string
          type: array
          title: Shipmentids
          description: Lista de IDs de despachos creados
      type: object
      required:
      - shipmentIds
      title: ShipmentCreateResponse
...
    ShipmentHistoryResponse:
      properties:
        shipmentId:
          type: string
          title: Shipmentid
        history:
          items:
            $ref: '#/components/schemas/ShipmentHistoryEvent'
          type: array
          title: History
      type: object
      required:
      - shipmentId
      - history
      title: ShipmentHistoryResponse
...
    ShipmentListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ShipmentListItem'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - data
      - pagination
      title: ShipmentListResponse
    ShipmentResponse:
      properties:
        shipmentId:
          type: string
          title: Shipmentid
        orderId:
          type: string
          title: Orderid
        customerName:
          type: string
          title: Customername
        address:
          type: string
          title: Address
        city:
          type: string
          title: City
        originCd:
          type: string
          title: Origincd
        volumetricWeight:
          type: number
          title: Volumetricweight
        shippingCost:
          type: integer
          title: Shippingcost
        weightKg:
          type: number
          title: Weightkg
        status:
          $ref: '#/components/schemas/ShipmentStatus'
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
        estimatedDelivery:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Estimateddelivery
      type: object
      required:
      - shipmentId
      - orderId
      - customerName
      - address
      - city
      - originCd
      - volumetricWeight
      - shippingCost
      - weightKg
      - status
      - createdAt
      - updatedAt
      title: ShipmentResponse
    ShipmentStatus:
      type: string
      enum:
      - PENDING
      - IN_TRANSIT
      - DELIVERED
      - CANCELLED
      - FAILED
      - RETURNED
      title: ShipmentStatus
    ShipmentUpdate:
      properties:
        status:
          $ref: '#/components/schemas/ShipmentStatus'
          description: Nuevo estado para la transición
      type: object
      required:
      - status
      title: ShipmentUpdate
...